fix default event
